Moved graphql API tests into a new package.
Renamed the base GraphQL API test to better reflect what it is specifically
for, and moved it and all its subclasses into a subpackage.

Also removed an unused method, just because.
