+++ backend/src/test/java/gov/cdc/usds/simplereport/api/graphql/ApiSmokeTest.java
@@ -1,4 +1,4 @@
-package gov.cdc.usds.simplereport.api;
+package gov.cdc.usds.simplereport.api.graphql;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -8,7 +8,7 @@ import java.io.IOException;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
 
-class ApiSmokeTest extends BaseApiTest {
+class ApiSmokeTest extends BaseGraphqlTest {
 
   @Test
   void smoketestPatientList() throws IOException {
